% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-dataset.R
\name{factorH_dataset}
\alias{factorH_dataset}
\alias{factorH-dataset}
\alias{dataset.factorH}
\title{Datasets in factorH}
\description{
Datasets in factorH
}
\details{
\subsection{What is in the example dataset?}{

mimicry is a real study on the \strong{chameleon effect} by Trzmielewska et
al. (2025) \doi{10.18290/rpsych2024.0019} about how \code{mimicry} vs other
movement conditions affect \code{liking} of an interlocutor. Potential
moderators include \code{gender} and \code{age} (with dichotomized age_cat, and a
3-level age_cat2). This makes it a natural playground for \code{multifactor}
rank-based analyses.

\if{html}{\out{<div class="sourceCode">}}\preformatted{table(mimicry$condition)
table(mimicry$gender)
table(mimicry$age_cat)
}\if{html}{\out{</div>}}
}
}
